refactor(ImageDetection): extract ClassificationAnnotations component
